structure rework
