+++ go.mod
@@ -1,15 +1,14 @@
 module gophertube
 
-go 1.23.0
+go 1.23.2
 
 toolchain go1.24.5
 
 require (
 	github.com/BurntSushi/toml v1.5.0
 	github.com/chzyer/readline v1.5.1
+	github.com/urfave/cli-altsrc/v3 v3.0.1
+	github.com/urfave/cli/v3 v3.3.8
 )
 
-require (
-	github.com/creack/pty v1.1.24 // indirect
-	golang.org/x/sys v0.34.0 // indirect
-)
+require golang.org/x/sys v0.34.0 // indirect
